fix: extract useCanvasMenuPositionSync composable to deduplicate context menu positioning logic (#9795)
